trusted-types/TrustedTypePolicyFactory-createPolicy.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js" ></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>
<body>
<script>
  function createHTMLTest(policy, expected, t) {
    let p = window.trustedTypes.createPolicy('SomeName', policy )
        .then(t.step_func_done(p => {
            assert_true(p.createHTML('whatever') instanceof TrustedHTML);
            assert_equals(p.createHTML('whatever') + "", expected);
    }));
  }

  async_test(t => {
    let policy = window.trustedTypes.createPolicy('SomeName', s => s )
        .then(t.step_func_done(policy => {
            assert_true(policy instanceof TrustedTypePolicy);
            assert_equals(policy.name, 'SomeName');
    }));
  }, "policy.name = name");

  async_test(t => {
    createHTMLTest( s => s, 'whatever', t);
  }, "trustedHTML.html = whatever");

  async_test(t => {
    createHTMLTest( s => null, "null", t);
  }, "trustedHTML.html = null");

  var str = 's';
  async_test(t => {
    createHTMLTest( s => str + s, str + 'whatever', t);
  }, "trustedHTML.html = whatever + global string");
/*
  async_test(t => {
    let policy = window.trustedTypes.createPolicy('SomeName', s => s )
        .then(t.step_func_done(policy => {
            assert_true(policy.createHTML('whatever') instanceof TrustedHTML);
            assert_equals(policy.createHTML('whatever') + "", 'whatever');
    }));
  }, "trustedHTML.html throws");*/
</script>
